cleanup useSubscription hook for solid-query
