+++ packages/solid-query/src/index.ts
@@ -1,0 +1,8 @@
+export type {
+  RspcQueryOptions,
+  RspcMutationOptions,
+  RspcSubscriptionOptions
+} from "./createOptionsProxy"
+export { createRSPCOptionsProxy } from "./createOptionsProxy"
+export type { SubscriptionStatus, SubscriptionResult } from "./useSubscription"
+export { useSubscription } from "./useSubscription"
